splitting out code actions and symbol info into their own files (#117)
